Updated README and renamed simulator -> track factory to better describe the operations
