+++ monitoring/rid_qualifier/README.md
@@ -1,10 +1,10 @@
 # RID Qualifier Tests
 
-This directory contains a series of tests for qualifying tests for Network Remote ID. It contains four main things:
+This directory contains a series of tests for qualifying Network Remote ID compliance. It contains four things:
 
-1. **Flight track dataset generator** (`flight_sim.py`): A API to generate flight paths and patterns within a bounding box. You can specify a bounding box in any part of the world the generator will create flight grid for the bounds and flight paths within that grid. The flights will follow a circular path and these are generated in the form of latitude / longitude within the bounding box.
+1. **Flight track dataset generator** (`flight_track_factory.py`): An API to generate multiple flight paths and patterns within a specified bounding box. You can specify a bounding box in any part of the world the generator will create a grid for flights for the bounds provided. In addition, circular flight paths will be generated within that grid. The flights will follow a circular path and these are generated in the form of latitude / longitude co-ordinates within the bounding box.
 
-2. **Position generator** (`flight_sim.py`): Once the flight tracks are generated, we simulate flights and "fly" them emitting position to be fed to the test harness, this is the `flight_sim` class. You can specify how many minutes (in seconds) you want the flights to "loop" over the flight paths. Specific details about the flights can be added as necessary. 
+2. **Test payload generator** (`position_generator.py`): Once the flight tracks are generated, we simulate flights and "fly" them emitting positions to be fed to the test harness, a file is generated that can be submitted by the test harness to the API. You can specify how many minutes (in seconds) you want the flights to "loop" over the flight paths. Specific details metadata about the flights are generated and added in preparation for submission to the API.
 
 3. **Test executor**: The test executor runs the test suite and generates flight paths and submits them to the test harness which in turn submits to the DSS.
